daily-laws-of-power/requirements.txt
google-cloud-storage==3.0.0
PyYAML
